+++ hollow/src/main/java/com/netflix/hollow/core/read/engine/list/HollowListTypeReadStateShard.java
@@ -66,7 +66,7 @@ class HollowListTypeReadStateShard {
         do {
             currentData = this.currentData;
 
-            long fixedLengthOffset = currentData.bitsPerListPointer * ordinal;
+            long fixedLengthOffset = (long)currentData.bitsPerListPointer * ordinal;
 
             long startAndEndElement = ordinal == 0 ?
                     currentData.listPointerArray.getElementValue(fixedLengthOffset, currentData.bitsPerListPointer) << currentData.bitsPerListPointer :
@@ -115,7 +115,7 @@ class HollowListTypeReadStateShard {
     }
 
     public long getApproximateHeapFootprintInBytes() {
-        long requiredListPointerBits = ((long)currentData.bitsPerListPointer * (currentData.maxOrdinal + 1));
+        long requiredListPointerBits = (currentData.bitsPerListPointer * ((long)currentData.maxOrdinal + 1));
         long requiredElementBits = (currentData.totalNumberOfElements * currentData.bitsPerElement);
         long requiredBits = requiredListPointerBits + requiredElementBits;
         return requiredBits / 8;
